Refactoring of compute statistics objects
This commit performs the following:

* Move logic to compute statistics objects for UNION ALL, PROJECT, SELECT,
GROUP BY and a LIMIT into a separate class that has static functions
that do the work. This remove the logic from CStatistics. The final goal
is to kill IStatistics and only have CStatistics that is a container for
statistical information such histograms, width, fempty, etc.

* Move the unittests for statistics on select predicates into its own test

*Get rid of CHistogramUtils and put these function in the CHistogram class

* Clean up some comments and remove dead code
