+++ CMakeLists.txt
@@ -6,7 +6,7 @@ project(gpopt LANGUAGES CXX C)
 set(CMAKE_CXX_STANDARD 98)
 set(GPORCA_VERSION_MAJOR 2)
 set(GPORCA_VERSION_MINOR 55)
-set(GPORCA_VERSION_PATCH 18)
+set(GPORCA_VERSION_PATCH 19)
 set(GPORCA_VERSION_STRING "${GPORCA_VERSION_MAJOR}.${GPORCA_VERSION_MINOR}.${GPORCA_VERSION_PATCH}")
 
 # Whenever an ABI-breaking change is made to GPORCA, this should be incremented.
