+++ libnaucrates/src/statistics/CFilterStatsProcessor.cpp
@@ -1,237 +1,170 @@
 //---------------------------------------------------------------------------
 //	Greenplum Database
-//	Copyright (C) 2017 Pivotal Inc.
+//	Copyright 2018 Pivotal, Inc.
 //
 //	@filename:
-//		CHistogramUtils.cpp
+//		CFilterStatsProcessor.cpp
 //
 //	@doc:
-//		Utility functions that operate on histograms
+//		Statistics helper routines for processing limit operations
 //---------------------------------------------------------------------------
 
-#include "naucrates/dxl/CDXLUtils.h"
-#include "naucrates/statistics/CHistogramUtils.h"
+#include "gpopt/operators/ops.h"
+#include "gpopt/optimizer/COptimizerConfig.h"
+
+#include "naucrates/statistics/CStatistics.h"
+#include "naucrates/statistics/CFilterStatsProcessor.h"
+#include "naucrates/statistics/CJoinStatsProcessor.h"
 #include "naucrates/statistics/CStatisticsUtils.h"
 #include "naucrates/statistics/CScaleFactorUtils.h"
 
-#include "gpos/common/CBitSet.h"
-#include "gpos/memory/CAutoMemoryPool.h"
-
-#include "gpopt/base/CColumnFactory.h"
-
-#include "gpopt/engine/CStatisticsConfig.h"
-#include "gpopt/optimizer/COptimizerConfig.h"
-
 using namespace gpopt;
 
-// append given histograms to current object
-void
-CHistogramUtils::AddHistograms
+// derive statistics for filter operation based on given scalar expression
+IStatistics *
+CFilterStatsProcessor::PstatsFilterForScalarExpr
 	(
 	IMemoryPool *pmp,
-	HMUlHist *phmulhistSrc,
-	HMUlHist *phmulhistDest
+	CExpressionHandle &exprhdl,
+	IStatistics *pstatsChild,
+	CExpression *pexprScalarLocal, // filter expression on local columns only
+	CExpression *pexprScalarOuterRefs, // filter expression involving outer references
+	DrgPstat *pdrgpstatOuter
 	)
 {
-	HMIterUlHist hmiterulhist(phmulhistSrc);
-	while (hmiterulhist.FAdvance())
+	GPOS_ASSERT(NULL != pstatsChild);
+	GPOS_ASSERT(NULL != pexprScalarLocal);
+	GPOS_ASSERT(NULL != pexprScalarOuterRefs);
+	GPOS_ASSERT(NULL != pdrgpstatOuter);
+
+	CColRefSet *pcrsOuterRefs = exprhdl.Pdprel()->PcrsOuter();
+
+	// TODO  June 13 2014, we currently only cap ndvs when we have a filter
+	// immediately on top of tables
+	BOOL fCapNdvs = (1 == exprhdl.Pdprel()->UlJoinDepth());
+
+	// extract local filter
+	CStatsPred *pstatspred = CStatsPredUtils::PstatspredExtract(pmp, pexprScalarLocal, pcrsOuterRefs);
+
+	// derive stats based on local filter
+	IStatistics *pstatsResult = CFilterStatsProcessor::PstatsFilter(pmp, dynamic_cast<CStatistics *>(pstatsChild), pstatspred, fCapNdvs);
+	pstatspred->Release();
+
+	if (exprhdl.FHasOuterRefs() && 0 < pdrgpstatOuter->UlLength())
 	{
-		ULONG ulColId = *(hmiterulhist.Pk());
-		const CHistogram *phist = hmiterulhist.Pt();
-		CStatisticsUtils::AddHistogram(pmp, ulColId, phist, phmulhistDest);
-	}
-}
-
-// add dummy histogram buckets and column information for the array of columns
-void
-CHistogramUtils::AddDummyHistogramAndWidthInfo
-	(
-	IMemoryPool *pmp,
-	CColumnFactory *pcf,
-	HMUlHist *phmulhistOutput,
-	HMUlDouble *phmuldoubleWidthOutput,
-	const DrgPul *pdrgpul,
-	BOOL fEmpty
-	)
-{
-	GPOS_ASSERT(NULL != pcf);
-	GPOS_ASSERT(NULL != phmulhistOutput);
-	GPOS_ASSERT(NULL != phmuldoubleWidthOutput);
-	GPOS_ASSERT(NULL != pdrgpul);
-
-	const ULONG ulCount = pdrgpul->UlLength();
-	// for computed aggregates, we're not going to be very smart right now
-	for (ULONG ul = 0; ul < ulCount; ul++)
-	{
-		ULONG ulColId = *(*pdrgpul)[ul];
-
-		CColRef *pcr = pcf->PcrLookup(ulColId);
-		GPOS_ASSERT(NULL != pcr);
-
-		CHistogram *phist = CHistogram::PhistDefault(pmp, pcr, fEmpty);
-		phmulhistOutput->FInsert(GPOS_NEW(pmp) ULONG(ulColId), phist);
-
-		CDouble dWidth = CStatisticsUtils::DDefaultColumnWidth(pcr->Pmdtype());
-		phmuldoubleWidthOutput->FInsert(GPOS_NEW(pmp) ULONG(ulColId), GPOS_NEW(pmp) CDouble(dWidth));
-	}
-}
-
-//	add empty histogram for the columns in the input histogram
-void
-CHistogramUtils::AddEmptyHistogram
-	(
-	IMemoryPool *pmp,
-	HMUlHist *phmulhistOutput,
-	HMUlHist *phmulhistInput
-	)
-{
-	GPOS_ASSERT(NULL != phmulhistOutput);
-	GPOS_ASSERT(NULL != phmulhistInput);
-
-	HMIterUlHist hmiterulhist(phmulhistInput);
-	while (hmiterulhist.FAdvance())
-	{
-		ULONG ulColId = *(hmiterulhist.Pk());
-
-		// empty histogram
-		CHistogram *phist =  GPOS_NEW(pmp) CHistogram(GPOS_NEW(pmp) DrgPbucket(pmp), false /* fWellDefined */);
-		phmulhistOutput->FInsert(GPOS_NEW(pmp) ULONG(ulColId), phist);
-	}
-}
-
-//	create a new histograms after applying the filter that is not
-//	an AND/OR predicate
-CHistogram *
-CHistogramUtils::PhistSimpleFilter
-	(
-	IMemoryPool *pmp,
-	CStatsPred *pstatspred,
-	CBitSet *pbsFilterColIds,
-	CHistogram *phistBefore,
-	CDouble *pdScaleFactorLast,
-	ULONG *pulColIdLast
-	)
-{
-	if (CStatsPred::EsptPoint == pstatspred->Espt())
-	{
-		CStatsPredPoint *pstatspredPoint = CStatsPredPoint::PstatspredConvert(pstatspred);
-		return PhistPointFilter(pmp, pstatspredPoint, pbsFilterColIds, phistBefore, pdScaleFactorLast, pulColIdLast);
+		// derive stats based on outer references
+		IStatistics *pstats = CJoinStatsProcessor::PstatsDeriveWithOuterRefs
+													(
+													pmp,
+													exprhdl,
+													pexprScalarOuterRefs,
+													pstatsResult,
+													pdrgpstatOuter,
+													IStatistics::EsjtInnerJoin
+													);
+		pstatsResult->Release();
+		pstatsResult = pstats;
 	}
 
-	if (CStatsPred::EsptLike == pstatspred->Espt())
-	{
-		CStatsPredLike *pstatspredLike = CStatsPredLike::PstatspredConvert(pstatspred);
+	return pstatsResult;
+}
 
-		return PhistLikeFilter(pmp, pstatspredLike, pbsFilterColIds, phistBefore, pdScaleFactorLast, pulColIdLast);
+// create new structure from a list of statistics filters
+CStatistics *
+CFilterStatsProcessor::PstatsFilter
+	(
+	IMemoryPool *pmp,
+	const CStatistics *pstatsInput,
+	CStatsPred *pstatspredBase,
+	BOOL fCapNdvs
+	)
+{
+	GPOS_ASSERT(NULL != pstatspredBase);
+
+	CDouble dRowsInput = std::max(CStatistics::DMinRows.DVal(), pstatsInput->DRows().DVal());
+	CDouble dScaleFactor(1.0);
+	ULONG ulNumPredicates = 1;
+	CDouble dRowsFilter = CStatistics::DMinRows;
+	HMUlHist *phmulhistNew = NULL;
+
+	HMUlHist *phmulhistCopy = pstatsInput->CopyHistograms(pmp);
+
+	CStatisticsConfig *pstatsconf = pstatsInput->PStatsConf();
+	if (pstatsInput->FEmpty())
+	{
+		phmulhistNew = GPOS_NEW(pmp) HMUlHist(pmp);
+		CHistogram::AddEmptyHistogram(pmp, phmulhistNew, phmulhistCopy);
+	}
+	else
+	{
+		if (CStatsPred::EsptDisj == pstatspredBase->Espt())
+		{
+			CStatsPredDisj *pstatspred = CStatsPredDisj::PstatspredConvert(pstatspredBase);
+
+			phmulhistNew  = PhmulhistApplyDisjFilter
+								(
+								pmp,
+								pstatsconf,
+								phmulhistCopy,
+								dRowsInput,
+								pstatspred,
+								&dScaleFactor
+								);
+		}
+		else
+		{
+			GPOS_ASSERT(CStatsPred::EsptConj == pstatspredBase->Espt());
+			CStatsPredConj *pstatspred = CStatsPredConj::PstatspredConvert(pstatspredBase);
+			ulNumPredicates = pstatspred->UlFilters();
+			phmulhistNew = PhmulhistApplyConjFilter
+							(
+							pmp,
+							pstatsconf,
+							phmulhistCopy,
+							dRowsInput,
+							pstatspred,
+							&dScaleFactor
+							);
+		}
+		GPOS_ASSERT(CStatistics::DMinRows.DVal() <= dScaleFactor.DVal());
+		dRowsFilter = dRowsInput / dScaleFactor;
+		dRowsFilter = std::max(CStatistics::DMinRows.DVal(), dRowsFilter.DVal());
 	}
 
-	CStatsPredUnsupported *pstatspredUnsupported = CStatsPredUnsupported::PstatspredConvert(pstatspred);
+	phmulhistCopy->Release();
 
-	return PhistUnsupportedPred(pmp, pstatspredUnsupported, pbsFilterColIds, phistBefore, pdScaleFactorLast, pulColIdLast);
-}
+	GPOS_ASSERT(dRowsFilter.DVal() <= dRowsInput.DVal());
 
-// create a new histograms after applying the point filter
-CHistogram *
-CHistogramUtils::PhistPointFilter
-	(
-	IMemoryPool *pmp,
-	CStatsPredPoint *pstatspred,
-	CBitSet *pbsFilterColIds,
-	CHistogram *phistBefore,
-	CDouble *pdScaleFactorLast,
-	ULONG *pulColIdLast
-	)
-{
-	GPOS_ASSERT(NULL != pstatspred);
-	GPOS_ASSERT(NULL != pbsFilterColIds);
-	GPOS_ASSERT(NULL != phistBefore);
+	if (fCapNdvs)
+	{
+		CStatistics::CapNDVs(dRowsFilter, phmulhistNew);
+	}
 
-	const ULONG ulColId = pstatspred->UlColId();
-	GPOS_ASSERT(CHistogram::FSupportsFilter(pstatspred->Escmpt()));
+	CStatistics *pstatsFilter = GPOS_NEW(pmp) CStatistics
+												(
+												pmp,
+												phmulhistNew,
+												pstatsInput->CopyWidths(pmp),
+												dRowsFilter,
+												pstatsInput->FEmpty(),
+												pstatsInput->UlNumberOfPredicates() + ulNumPredicates
+												);
 
-	CPoint *ppoint = pstatspred->Ppoint();
+	// since the filter operation is reductive, we choose the bounding method that takes
+	// the minimum of the cardinality upper bound of the source column (in the input hash map)
+	// and estimated output cardinality
+	CStatisticsUtils::ComputeCardUpperBounds(pmp, pstatsInput, pstatsFilter, dRowsFilter, CStatistics::EcbmMin /* ecbm */);
 
-	// note column id
-	(void) pbsFilterColIds->FExchangeSet(ulColId);
-
-	CDouble dScaleFactorLocal(1.0);
-	CHistogram *phistAfter = phistBefore->PhistFilterNormalized(pmp, pstatspred->Escmpt(), ppoint, &dScaleFactorLocal);
-
-	GPOS_ASSERT(DOUBLE(1.0) <= dScaleFactorLocal.DVal());
-
-	*pdScaleFactorLast = *pdScaleFactorLast * dScaleFactorLocal;
-	*pulColIdLast = ulColId;
-
-	return phistAfter;
-}
-
-
-//	create a new histograms for an unsupported predicate
-CHistogram *
-CHistogramUtils::PhistUnsupportedPred
-	(
-	IMemoryPool *pmp,
-	CStatsPredUnsupported *pstatspred,
-	CBitSet *pbsFilterColIds,
-	CHistogram *phistBefore,
-	CDouble *pdScaleFactorLast,
-	ULONG *pulColIdLast
-	)
-{
-	GPOS_ASSERT(NULL != pstatspred);
-	GPOS_ASSERT(NULL != pbsFilterColIds);
-	GPOS_ASSERT(NULL != phistBefore);
-
-	const ULONG ulColId = pstatspred->UlColId();
-
-	// note column id
-	(void) pbsFilterColIds->FExchangeSet(ulColId);
-
-	// generate after histogram
-	CHistogram *phistAfter = phistBefore->PhistCopy(pmp);
-	GPOS_ASSERT(NULL != phistAfter);
-
-	*pdScaleFactorLast = *pdScaleFactorLast * pstatspred->DScaleFactor();
-	*pulColIdLast = ulColId;
-
-	return phistAfter;
-}
-
-//	create a new histograms after applying the LIKE filter
-CHistogram *
-CHistogramUtils::PhistLikeFilter
-	(
-	IMemoryPool *pmp,
-	CStatsPredLike *pstatspred,
-	CBitSet *pbsFilterColIds,
-	CHistogram *phistBefore,
-	CDouble *pdScaleFactorLast,
-	ULONG *pulColIdLast
-	)
-{
-	GPOS_ASSERT(NULL != pstatspred);
-	GPOS_ASSERT(NULL != pbsFilterColIds);
-	GPOS_ASSERT(NULL != phistBefore);
-
-	const ULONG ulColId = pstatspred->UlColId();
-
-	// note column id
-	(void) pbsFilterColIds->FExchangeSet(ulColId);
-	CHistogram *phistAfter = phistBefore->PhistCopy(pmp);
-
-	*pdScaleFactorLast = *pdScaleFactorLast * pstatspred->DDefaultScaleFactor();
-	*pulColIdLast = ulColId;
-
-	return phistAfter;
+	return pstatsFilter;
 }
 
 // create a new hash map of histograms after applying a conjunctive
 // or a disjunctive filter
 HMUlHist *
-CHistogramUtils::PhmulhistApplyConjOrDisjFilter
+CFilterStatsProcessor::PhmulhistApplyConjOrDisjFilter
 	(
 	IMemoryPool *pmp,
-	CStatisticsConfig *pstatsconf,
+	const CStatisticsConfig *pstatsconf,
 	HMUlHist *phmulhistInput,
 	CDouble dRowsInput,
 	CStatsPred *pstatspred,
@@ -276,10 +209,10 @@ CHistogramUtils::PhmulhistApplyConjOrDisjFilter
 
 // create new hash map of histograms after applying conjunctive predicates
 HMUlHist *
-CHistogramUtils::PhmulhistApplyConjFilter
+CFilterStatsProcessor::PhmulhistApplyConjFilter
 	(
 	IMemoryPool *pmp,
-	CStatisticsConfig *pstatsconf,
+	const CStatisticsConfig *pstatsconf,
 	HMUlHist *phmulhistInput,
 	CDouble dRowsInput,
 	CStatsPredConj *pstatspredConj,
@@ -337,7 +270,7 @@ CHistogramUtils::PhmulhistApplyConjFilter
 			GPOS_ASSERT(NULL != phistBefore);
 
 			CHistogram *phistResult = NULL;
-			phistResult = CHistogramUtils::PhistSimpleFilter(pmp, pstatspredChild, pbsFilterColIds, phistBefore, &dScaleFactorLast, &ulColIdLast);
+			phistResult = PhistSimpleFilter(pmp, pstatspredChild, pbsFilterColIds, phistBefore, &dScaleFactorLast, &ulColIdLast);
 			GPOS_DELETE(phistBefore);
 
 			GPOS_ASSERT(NULL != phistResult);
@@ -425,10 +358,10 @@ CHistogramUtils::PhmulhistApplyConjFilter
 
 // create new hash map of histograms after applying disjunctive predicates
 HMUlHist *
-CHistogramUtils::PhmulhistApplyDisjFilter
+CFilterStatsProcessor::PhmulhistApplyDisjFilter
 	(
 	IMemoryPool *pmp,
-	CStatisticsConfig *pstatsconf,
+	const CStatisticsConfig *pstatsconf,
 	HMUlHist *phmulhistInput,
 	CDouble dRowsInput,
 	CStatsPredDisj *pstatspredDisj,
@@ -498,7 +431,7 @@ CHistogramUtils::PhmulhistApplyDisjFilter
 		if (fPredSimple)
 		{
 			GPOS_ASSERT(NULL != phist);
-			phistDisjChildCol = CHistogramUtils::PhistSimpleFilter(pmp, pstatspredChild, pbsFilterColIds, phist, &dScaleFactorChild, &ulColIdPrev);
+			phistDisjChildCol = PhistSimpleFilter(pmp, pstatspredChild, pbsFilterColIds, phist, &dScaleFactorChild, &ulColIdPrev);
 
 			CHistogram *phistInput = phmulhistInput->PtLookup(&ulColId);
 			GPOS_ASSERT(NULL != phistInput);
@@ -511,15 +444,15 @@ CHistogramUtils::PhmulhistApplyDisjFilter
 		}
 		else
 		{
-			phmulhistChild = CHistogramUtils::PhmulhistApplyConjOrDisjFilter
-								  	  			(
-								  	  			pmp,
-												pstatsconf,
-												phmulhistInput,
-												dRowsInput,
-												pstatspredChild,
-												&dScaleFactorChild
-								  	  			);
+			phmulhistChild = PhmulhistApplyConjOrDisjFilter
+								(
+								pmp,
+								pstatsconf,
+								phmulhistInput,
+								dRowsInput,
+								pstatspredChild,
+								&dScaleFactorChild
+								);
 
 			GPOS_ASSERT_IMP(CStatsPred::EsptDisj == pstatspredChild->Espt(), ULONG_MAX != ulColId);
 
@@ -602,7 +535,7 @@ CHistogramUtils::PhmulhistApplyDisjFilter
 
 	*pdScaleFactor = CScaleFactorUtils::DScaleFactorCumulativeDisj(pstatsconf, pdrgpdScaleFactor, dRowsInput);
 
-	AddHistograms(pmp, phmulhistInput, phmulhistResultDisj);
+	CHistogram::AddHistograms(pmp, phmulhistInput, phmulhistResultDisj);
 
 	pbsStatsNonUpdateableCols->Release();
 
@@ -613,52 +546,135 @@ CHistogramUtils::PhmulhistApplyDisjFilter
 	return phmulhistResultDisj;
 }
 
-// check if the join cardinality estimation can be done based on NDV alone
-BOOL
-CHistogramUtils::FNDVBasedCardEstimation
+//	create a new histograms after applying the filter that is not
+//	an AND/OR predicate
+CHistogram *
+CFilterStatsProcessor::PhistSimpleFilter
 	(
-	const CHistogram *phist
+	IMemoryPool *pmp,
+	CStatsPred *pstatspred,
+	CBitSet *pbsFilterColIds,
+	CHistogram *phistBefore,
+	CDouble *pdScaleFactorLast,
+	ULONG *pulColIdLast
 	)
 {
-	GPOS_ASSERT(NULL != phist);
-
-	if (0 == phist->UlBuckets())
+	if (CStatsPred::EsptPoint == pstatspred->Espt())
 	{
-		// no buckets, so join cardinality estimation is based solely on NDV remain
-		return true;
+		CStatsPredPoint *pstatspredPoint = CStatsPredPoint::PstatspredConvert(pstatspred);
+		return PhistPointFilter(pmp, pstatspredPoint, pbsFilterColIds, phistBefore, pdScaleFactorLast, pulColIdLast);
 	}
 
-	const IBucket *pbucket = (*phist->Pdrgpbucket())[0];
-
-	IDatum *pdatum = pbucket->PpLower()->Pdatum();
-
-	IMDType::ETypeInfo eti = pdatum->Eti();
-	if (IMDType::EtiInt2 == eti ||
-		IMDType::EtiInt4 == eti ||
-		IMDType::EtiInt8 == eti ||
-		IMDType::EtiBool == eti ||
-		IMDType::EtiOid == eti )
+	if (CStatsPred::EsptLike == pstatspred->Espt())
 	{
-		return false;
+		CStatsPredLike *pstatspredLike = CStatsPredLike::PstatspredConvert(pstatspred);
+
+		return PhistLikeFilter(pmp, pstatspredLike, pbsFilterColIds, phistBefore, pdScaleFactorLast, pulColIdLast);
 	}
 
-	BOOL fRes = true;
-	if (pdatum->FStatsMappable())
-	{
-		IDatumStatisticsMappable *pdatumMappable = (IDatumStatisticsMappable *) pdatum;
+	CStatsPredUnsupported *pstatspredUnsupported = CStatsPredUnsupported::PstatspredConvert(pstatspred);
 
-		if (pdatumMappable->FHasStatsDoubleMapping())
-		{
-			fRes = false;
-		}
-	}
+	return PhistUnsupportedPred(pmp, pstatspredUnsupported, pbsFilterColIds, phistBefore, pdScaleFactorLast, pulColIdLast);
+}
 
-	return fRes;
+// create a new histograms after applying the point filter
+CHistogram *
+CFilterStatsProcessor::PhistPointFilter
+	(
+	IMemoryPool *pmp,
+	CStatsPredPoint *pstatspred,
+	CBitSet *pbsFilterColIds,
+	CHistogram *phistBefore,
+	CDouble *pdScaleFactorLast,
+	ULONG *pulColIdLast
+	)
+{
+	GPOS_ASSERT(NULL != pstatspred);
+	GPOS_ASSERT(NULL != pbsFilterColIds);
+	GPOS_ASSERT(NULL != phistBefore);
+
+	const ULONG ulColId = pstatspred->UlColId();
+	GPOS_ASSERT(CHistogram::FSupportsFilter(pstatspred->Escmpt()));
+
+	CPoint *ppoint = pstatspred->Ppoint();
+
+	// note column id
+	(void) pbsFilterColIds->FExchangeSet(ulColId);
+
+	CDouble dScaleFactorLocal(1.0);
+	CHistogram *phistAfter = phistBefore->PhistFilterNormalized(pmp, pstatspred->Escmpt(), ppoint, &dScaleFactorLocal);
+
+	GPOS_ASSERT(DOUBLE(1.0) <= dScaleFactorLocal.DVal());
+
+	*pdScaleFactorLast = *pdScaleFactorLast * dScaleFactorLocal;
+	*pulColIdLast = ulColId;
+
+	return phistAfter;
+}
+
+
+//	create a new histograms for an unsupported predicate
+CHistogram *
+CFilterStatsProcessor::PhistUnsupportedPred
+	(
+	IMemoryPool *pmp,
+	CStatsPredUnsupported *pstatspred,
+	CBitSet *pbsFilterColIds,
+	CHistogram *phistBefore,
+	CDouble *pdScaleFactorLast,
+	ULONG *pulColIdLast
+	)
+{
+	GPOS_ASSERT(NULL != pstatspred);
+	GPOS_ASSERT(NULL != pbsFilterColIds);
+	GPOS_ASSERT(NULL != phistBefore);
+
+	const ULONG ulColId = pstatspred->UlColId();
+
+	// note column id
+	(void) pbsFilterColIds->FExchangeSet(ulColId);
+
+	// generate after histogram
+	CHistogram *phistAfter = phistBefore->PhistCopy(pmp);
+	GPOS_ASSERT(NULL != phistAfter);
+
+	*pdScaleFactorLast = *pdScaleFactorLast * pstatspred->DScaleFactor();
+	*pulColIdLast = ulColId;
+
+	return phistAfter;
+}
+
+//	create a new histograms after applying the LIKE filter
+CHistogram *
+CFilterStatsProcessor::PhistLikeFilter
+	(
+	IMemoryPool *pmp,
+	CStatsPredLike *pstatspred,
+	CBitSet *pbsFilterColIds,
+	CHistogram *phistBefore,
+	CDouble *pdScaleFactorLast,
+	ULONG *pulColIdLast
+	)
+{
+	GPOS_ASSERT(NULL != pstatspred);
+	GPOS_ASSERT(NULL != pbsFilterColIds);
+	GPOS_ASSERT(NULL != phistBefore);
+
+	const ULONG ulColId = pstatspred->UlColId();
+
+	// note column id
+	(void) pbsFilterColIds->FExchangeSet(ulColId);
+	CHistogram *phistAfter = phistBefore->PhistCopy(pmp);
+
+	*pdScaleFactorLast = *pdScaleFactorLast * pstatspred->DDefaultScaleFactor();
+	*pulColIdLast = ulColId;
+
+	return phistAfter;
 }
 
 // check if the column is a new column for statistic calculation
 BOOL
-CHistogramUtils::FNewStatsColumn
+CFilterStatsProcessor::FNewStatsColumn
 	(
 	ULONG ulColId,
 	ULONG ulColIdLast
@@ -668,4 +684,3 @@ CHistogramUtils::FNewStatsColumn
 }
 
 // EOF
-
