+++ libnaucrates/src/statistics/CStatisticsUtils.cpp
@@ -24,6 +24,7 @@
 
 #include "naucrates/statistics/CStatisticsUtils.h"
 #include "naucrates/statistics/CJoinStatsProcessor.h"
+#include "naucrates/statistics/CFilterStatsProcessor.h"
 #include "naucrates/statistics/CStatistics.h"
 #include "naucrates/statistics/CStatsPredUtils.h"
 #include "naucrates/statistics/CStatsPredDisj.h"
@@ -1100,69 +1101,6 @@ CStatisticsUtils::PdatumNull
 	return pdatum;
 }
 
-
-
-
-//---------------------------------------------------------------------------
-//	@function:
-//		CStatisticsUtils::PstatsFilter
-//
-//	@doc:
-//		Derive statistics for filter operation based on given scalar expression
-//
-//---------------------------------------------------------------------------
-IStatistics *
-CStatisticsUtils::PstatsFilter
-	(
-	IMemoryPool *pmp,
-	CExpressionHandle &exprhdl,
-	IStatistics *pstatsChild,
-	CExpression *pexprScalarLocal, // filter expression on local columns only
-	CExpression *pexprScalarOuterRefs, // filter expression involving outer references
-	DrgPstat *pdrgpstatOuter
-	)
-{
-	GPOS_ASSERT(NULL != pstatsChild);
-	GPOS_ASSERT(NULL != pexprScalarLocal);
-	GPOS_ASSERT(NULL != pexprScalarOuterRefs);
-	GPOS_ASSERT(NULL != pdrgpstatOuter);
-
-	CColRefSet *pcrsOuterRefs = exprhdl.Pdprel()->PcrsOuter();
-	
-	// TODO  June 13 2014, we currently only cap ndvs when we have a filter
-	// immediately on top of tables
-	BOOL fCapNdvs = (1 == exprhdl.Pdprel()->UlJoinDepth());
-
-	// extract local filter
-	CStatsPred *pstatspred = CStatsPredUtils::PstatspredExtract(pmp, pexprScalarLocal, pcrsOuterRefs);
-
-	// derive stats based on local filter
-	IStatistics *pstatsResult = pstatsChild->PstatsFilter(pmp, pstatspred, fCapNdvs);
-	pstatspred->Release();
-
-	if (exprhdl.FHasOuterRefs() && 0 < pdrgpstatOuter->UlLength())
-	{
-		// derive stats based on outer references
-		IStatistics *pstats = CJoinStatsProcessor::PstatsDeriveWithOuterRefs(pmp,
-																				 exprhdl,
-																				 pexprScalarOuterRefs,
-																				 pstatsResult,
-																				 pdrgpstatOuter,
-																				 IStatistics::EsjtInnerJoin
-																				);
-		pstatsResult->Release();
-		pstatsResult = pstats;
-	}
-
-	return pstatsResult;
-}
-
-
-
-
-
-
-
 //---------------------------------------------------------------------------
 //	@function:
 //		CStatisticsUtils::PstatsDynamicScan
@@ -1297,7 +1235,7 @@ CStatisticsUtils::PstatsIndexGet
 	IStatistics *pstatsBaseTable = CLogical::PstatsBaseTable(pmp, exprhdl, ptabdesc, pcrsUsed);
 	pcrsUsed->Release();
 
-	IStatistics *pstats = PstatsFilter(pmp, exprhdl, pstatsBaseTable, pexprLocal, pexprOuterRefs, pdrgpstatCtxt);
+	IStatistics *pstats = CFilterStatsProcessor::PstatsFilterForScalarExpr(pmp, exprhdl, pstatsBaseTable, pexprLocal, pexprOuterRefs, pdrgpstatCtxt);
 
 	pstatsBaseTable->Release();
 	pexprLocal->Release();
@@ -1342,7 +1280,7 @@ CStatisticsUtils::PstatsBitmapTableGet
 	pcrsUsed->Difference(pcrsOuter);
 	IStatistics *pstatsBaseTable = CLogical::PstatsBaseTable(pmp, exprhdl, ptabdesc, pcrsUsed);
 	pcrsUsed->Release();
-	IStatistics *pstats = PstatsFilter
+	IStatistics *pstats = CFilterStatsProcessor::PstatsFilterForScalarExpr
 							(
 							pmp,
 							exprhdl,
@@ -1476,7 +1414,7 @@ DrgPdouble *
 CStatisticsUtils::PdrgPdoubleNDV
 	(
 	IMemoryPool *pmp,
-	CStatisticsConfig *pstatsconf,
+	const CStatisticsConfig *pstatsconf,
 	const IStatistics *pstats,
 	CColRefSet *pcrsGrpCols,
 	CBitSet *pbsKeys // keys derived during optimization
@@ -1605,7 +1543,7 @@ CDouble
 CStatisticsUtils::DMaxGroupsFromSource
 	(
 	IMemoryPool *pmp,
-	CStatisticsConfig *pstatsconf,
+	const CStatisticsConfig *pstatsconf,
 	CStatistics *pstatsInput,
 	const DrgPul *pdrgpulPerSrc
 	)
@@ -1658,7 +1596,7 @@ CStatisticsUtils::DGroups
 	(
 	IMemoryPool *pmp,
 	IStatistics *pstats,
-	CStatisticsConfig *pstatsconf,
+	const CStatisticsConfig *pstatsconf,
 	DrgPul *pdrgpulGC,
 	CBitSet *pbsKeys // keys derived during optimization
 	)
@@ -1701,7 +1639,7 @@ CStatisticsUtils::DGroups
 CDouble
 CStatisticsUtils::DNumOfDistinctVal
 	(
-	CStatisticsConfig *pstatsconf,
+	const CStatisticsConfig *pstatsconf,
 	DrgPdouble *pdrgpdNDV
 	)
 {
