+++ server/include/unittest/dxl/statistics/CStatisticsTest.h
@@ -41,32 +41,7 @@ namespace gpnaucrates
 	//---------------------------------------------------------------------------
 	class CStatisticsTest
 	{
-
-		// shorthand for functions for generating the disjunctive filter predicates
-		typedef CStatsPred *(FnPstatspredDisj)(IMemoryPool *pmp);
-
 		private:
-
-			// triplet consisting of comparison type, double value and its byte array representation
-			struct SStatsCmpValElem
-			{
-				CStatsPred::EStatsCmpType m_escmpt; // comparison operator
-				const WCHAR *m_wsz; // byte array representation
-				CDouble m_dVal; // double value
-			}; // SStatsCmpValElem
-
-			// test case for disjunctive filter evaluation
-			struct SStatsFilterSTestCase
-			{
-				// input stats dxl file
-				const CHAR *m_szInputFile;
-
-				// output stats dxl file
-				const CHAR *m_szOutputFile;
-
-				// filter predicate generation function pointer
-				FnPstatspredDisj *m_pf;
-			}; // SStatsFilterSTestCase
 
 			// test case for union all evaluation
 			struct SStatsUnionAllSTestCase
@@ -117,10 +92,6 @@ namespace gpnaucrates
 			static
 			CHistogram* PhistExampleInt4Dim(IMemoryPool *pmp);
 
-			// create a numeric predicate on a particular column
-			static
-			DrgPstatspred *PdrgppredfilterNumeric(IMemoryPool *pmp, ULONG ulColId, SStatsCmpValElem statsCmpValElem);
-
 			// helper function that generates an array of ULONG pointers
 			static
 			DrgPul *Pdrgpul
@@ -141,10 +112,6 @@ namespace gpnaucrates
 				return pdrgpul;
 			}
 
- 			// helper method to iterate over an array generated filter predicates for stats evaluation
-			static
-			GPOS_RESULT EresUnittest_CStatistics(SStatsFilterSTestCase rgstatsdisjtc[], ULONG ulTestCases);
-
 			// create a table descriptor with two columns having the given names
 			static
 			CTableDescriptor *PtabdescTwoColumnSource
@@ -158,17 +125,6 @@ namespace gpnaucrates
 
 		public:
 
-			// generate an array of filter given a column identifier, comparison type, and array of integer point
-			static
-			DrgPstatspred *PdrgpstatspredInteger
-				(
-				IMemoryPool *pmp,
-				ULONG ulColId,
-				CStatsPred::EStatsCmpType escmpt,
-				INT *piVals,
-				ULONG ulVals
-				);
-
 			// example filter
 			static
 			DrgPstatspred *Pdrgpstatspred1(IMemoryPool *pmp);
@@ -176,108 +132,9 @@ namespace gpnaucrates
 			static
 			DrgPstatspred *Pdrgpstatspred2(IMemoryPool *pmp);
 
-			// create a filter on a column with null values
-			static
-			CStatsPred *PstatspredNullableCols(IMemoryPool *pmp);
-
-			// create a point filter where the constant is null
-			static
-			CStatsPred *PstatspredWithNullConstant(IMemoryPool *pmp);
-
-			// create a 'is not null' point filter
-			static
-			CStatsPred *PstatspredNotNull(IMemoryPool *pmp);
-
-			// nested AND and OR predicates
-			static
-			CStatsPred *PstatspredNestedPredDiffCol1(IMemoryPool *pmp);
-
-			// nested AND and OR predicates
-			static
-			CStatsPred *PstatspredNestedPredDiffCol2(IMemoryPool *pmp);
-
-			// nested AND and OR predicates
-			static
-			CStatsPred *PstatspredNestedPredCommonCol1(IMemoryPool *pmp);
-
-			// nested AND and OR predicates
-			static
-			CStatsPred *PstatspredNestedPredCommonCol2(IMemoryPool *pmp);
-
-			// nested AND and OR predicates
-			static
-			CStatsPred *PstatspredNestedSharedCol(IMemoryPool *pmp);
-
-			// nested AND and OR predicates
-			static
-			CStatsPred *PstatspredDisjOverConjSameCol1(IMemoryPool *pmp);
-
-			// nested AND and OR predicates
-			static
-			CStatsPred *PstatspredDisjOverConjSameCol2(IMemoryPool *pmp);
-
-			// nested AND and OR predicates
-			static
-			CStatsPred *PstatspredDisjOverConjSameCol3(IMemoryPool *pmp);
-
-			// nested AND and OR predicates
-			static
-			CStatsPred *PstatspredDisjOverConjSameCol4(IMemoryPool *pmp);
-
-			// conjunctive predicates
-			static
-			CStatsPred *PstatspredConj(IMemoryPool *pmp);
-
-			// nested AND and OR predicates
-			static
-			CStatsPred *PstatspredDisjOverConjDifferentCol1(IMemoryPool *pmp);
-
-			static
-			CStatsPred *PstatspredDisjOverConjMultipleIdenticalCols(IMemoryPool *pmp);
-
-			// disjunction filters
-			static
-			CStatsPred *PstatspredDisj1(IMemoryPool *pmp);
-			
-			// disjunction filters
-			static
-			CStatsPred *PstatspredDisj2(IMemoryPool *pmp);
-
-			// disjunction filters
-			static
-			CStatsPred *PstatspredDisj3(IMemoryPool *pmp);
-
-			// disjunction filters
-			static
-			CStatsPred *PstatspredDisj4(IMemoryPool *pmp);
-
-			// disjunction filters
-			static
-			CStatsPred *PstatspredDisj5(IMemoryPool *pmp);
-
-			// disjunction filters
-			static
-			CStatsPred *PstatspredDisj6(IMemoryPool *pmp);
-
-			// disjunction filters
-			static
-			CStatsPred *PstatspredDisj7(IMemoryPool *pmp);
-
-			// disjunction filters
-			static
-			CStatsPred *PstatspredDisj8(IMemoryPool *pmp);
-
 			// unittests
 			static
 			GPOS_RESULT EresUnittest();
-
-			// join of histograms with NDVRemain information
-			static
-			GPOS_RESULT EresUnittest_JoinNDVRemain();
-
-			// join buckets tests
-			static
-			GPOS_RESULT EresUnittest_Join();
 
 			// union all tests
 			static
@@ -287,49 +144,9 @@ namespace gpnaucrates
 			static
 			GPOS_RESULT EresUnittest_CStatisticsBasic();
 
-			// basic statistics parsing
-			static
-			GPOS_RESULT EresUnittest_CStatisticsBasicsFromDXL();
-
-			// testing select predicates
-			static
-			GPOS_RESULT EresUnittest_CStatisticsFilter();
-
-			// testing nested AND / OR predicates
-			static
-			GPOS_RESULT EresUnittest_CStatisticsNestedPred();
-
-			// test disjunctive filter
-			static
-			GPOS_RESULT EresUnittest_CStatisticsFilterDisj();
-
-			// test conjunctive filter
-			static
-			GPOS_RESULT EresUnittest_CStatisticsFilterConj();
-
-			// DXL based test on numeric data types
-			static
-			GPOS_RESULT EresUnittest_CStatisticsBasicsFromDXLNumeric();
-
-			// compare the derived statistics with the statistics in the outputfile
-			static
-			GPOS_RESULT EresUnittest_CStatisticsCompare
-				(
-				IMemoryPool *pmp,
-				CMDAccessor *pmda,
-				DrgPstats *pdrgpstatBefore,
-				CStatsPred *pstatspred,
-				const CHAR *szDXLOutput,
-				BOOL fApplyTwice = false
-				);
-
 			// exercise stats derivation during optimization
 			static
 			GPOS_RESULT EresUnittest_CStatisticsSelectDerivation();
-
-			// test for accumulating cardinality in disjunctive and conjunctive predicates
-	 		static
-			GPOS_RESULT EresUnittest_CStatisticsAccumulateCard();
 
 			// GbAgg test when grouping on repeated columns
 			static
